+++ CHANGELOG.md
@@ -5,6 +5,10 @@ format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 #### Added
+- In the Skill List activity, add support for displaying skills allowed
+  by the Dilettante skill
+  ([issue #23](https://github.com/kuhrusty/MorbadScorepad/issues/23)).
+
 #### Changed